requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
oauthlib==3.1.0
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy~=3.10.0
urllib3==1.26.5
python-dotenv~=0.18.0